+++ packages/build-field-types/__fixtures__/basic-field-type/src/index.js
@@ -1,4 +1,4 @@
-import { importView } from '@keystone-alpha/build-field-types';
+import { importView } from '@keystone/build-field-types';
 
 export let MyCoolFieldType = {
   views: {