+++ packages/fields/src/types/DateTime/index.js
@@ -1,5 +1,5 @@
 import { DateTime, MongoDateTimeInterface, KnexDateTimeInterface } from './Implementation';
-import { importView } from '@keystone-alpha/build-field-types';
+import { importView } from '@keystone/build-field-types';
 
 export default {
   type: 'DateTime',