+++ packages/fields/src/types/Decimal/index.js
@@ -1,5 +1,5 @@
 import { Decimal, MongoDecimalInterface, KnexDecimalInterface } from './Implementation';
-import { importView } from '@keystone-alpha/build-field-types';
+import { importView } from '@keystone/build-field-types';
 
 export default {
   type: 'Decimal',