+++ packages/fields/src/types/Select/index.js
@@ -1,5 +1,5 @@
 import { Select, MongoSelectInterface, KnexSelectInterface } from './Implementation';
-import { importView } from '@keystone-alpha/build-field-types';
+import { importView } from '@keystone/build-field-types';
 
 export default {
   type: 'Select',